Updated version 4.8.0

+++ ClanSys/commands/admins/moderation/clear.js
@@ -1,7 +1,15 @@
 
-const { ActionRowBuilder, ButtonBuilder, EmbedBuilder, PermissionsBitField, SlashCommandBuilder } = require('discord.js');
+// Require and set
+const Discord = require('discord.js');
+const { PermissionsBitField, EmbedBuilder, SlashCommandBuilder } = Discord;
+const { DateTime } = require('luxon');
+const timeFormat = 'LL'+'/'+'dd'+'/'+'yyyy'+'-'+'h'+':'+'mm'+':'+'ss'+'-'+'a';
 require('dotenv').config();
+
 module.exports = {
+	cooldown: 5,
+	admin: 'true',
+	nsfw: 'false',
     data: new SlashCommandBuilder()
         .setName('clear')
         .setDescription('clear up to 1000 messages of the past two weeks')
@@ -10,7 +18,7 @@ module.exports = {
             PermissionsBitField.Flags.ViewAuditLog
             | PermissionsBitField.Flags.KickMembers
             | PermissionsBitField.Flags.ManageChannels
-            | PermissionsBitField.Flags.ManageEmojisAndStickers
+            | PermissionsBitField.Flags.ManageGuildExpressions
             | PermissionsBitField.Flags.ManageGuild
             | PermissionsBitField.Flags.ManageMessages
             | PermissionsBitField.Flags.ManageRoles
@@ -33,8 +41,6 @@ module.exports = {
                 .setRequired(false),
         )
     ,
-    nsfw: 'false',       // NSFW variable = 'true', No NSFW variable = 'false'.
-    admin: 'true',      // Admin Command = 'true', No Admin Command = 'false'.
     async execute(interaction) {
         if (interaction != null || interaction.channel.id != null || interaction.guild.id != null) {
             // SQLite
@@ -55,41 +61,49 @@ module.exports = {
             if (dataCommandAdmin == null) { dataCommandAdmin = { Clear: 'true' }; };
             if (dataChannellog == null) {console.log('No logging Channel in database');return;};
             // Context
+
+            let lang = require(`../../../.${dataLang.Lang}`);
             if (dataCommandAdmin.Clear === 'true') {
-                const stringAmount = interaction.options.getInteger('amount');
-                const stringUser = interaction.options.getUser('member');
-                let guild = await interaction.client.guilds.fetch(getGuildID);
-                let channel = await guild.channels.fetch(getChannelID);
-                if (stringAmount > 100) {
-                    await interaction.reply({ content: 'Sorry I only can do 100 at a time.', ephemeral: true })
-                } else
-                if (stringAmount <= 100) {
-                    let messages  = await channel.messages.fetch({ limit: stringAmount });
-                    let embedMsg = `${messages.size} Messages from the past Two weeks got cleared.`
-                    // Delete user messages
-                    if (stringUser) {
-                        messages = await channel.messages.fetch({ limit: 100 })
-                        messages = messages.filter((m) => m.author.id == stringUser.id);
-                        messages = messages.map(function(obj){return obj;});
-                        messages = messages.slice(0, stringAmount);
-                        embedMsg = `${messages.length} Messages from ${stringUser} from the past Two weeks got cleared.`;
+                let permissions = interaction.member.permissions;
+                if (permissions.has(PermissionsBitField.Flags.ViewAuditLog) || permissions.has(PermissionsBitField.Flags.ManageChannels)) {
+                    const stringAmount = interaction.options.getInteger('amount');
+                    const stringUser = interaction.options.getUser('member');
+                    let guild = await interaction.client.guilds.fetch(getGuildID);
+                    let channel = await guild.channels.fetch(getChannelID);
+                    if (stringAmount > 100) {
+                        await interaction.reply({ content: `${lang.admin.clear.max100}`, ephemeral: true })
+                    } else
+                    if (stringAmount <= 100) {
+                        let messages  = await channel.messages.fetch({ limit: stringAmount });
+                        let embedMsg = `${messages.size} ${lang.admin.clear.msgclear1}`
+                        // Delete user messages
+                        if (stringUser) {
+                            messages = await channel.messages.fetch({ limit: 100 })
+                            messages = messages.filter((m) => m.author.id == stringUser.id);
+                            messages = messages.map(function(obj){return obj;});
+                            messages = messages.slice(0, stringAmount);
+                            embedMsg = `${messages.length} ${lang.admin.clear.msgfrom} ${stringUser} ${lang.admin.clear.msgclear2}`;
+                        };
+                        // Delete the message
+                        await channel.bulkDelete(messages, true);
+                        const embed = new EmbedBuilder()
+                            .addFields(
+                                { name: `${lang.admin.clear.clearedtitle}`, value: `${embedMsg}` }
+                            )
+                        await interaction.reply({ embeds: [embed], ephemeral: true });
+                        // let embedLogMsg = `${messages.length} ${lang.admin.clear.msgfrom} ${stringUser} ${lang.admin.clear.msgclear2} ${channel}.`
+                        // const embedLog = new EmbedBuilder()
+                        //     .addFields(
+                        //         { name: `${lang.admin.clear.clearedtitle}`, value: `${embedLogMsg}` }
+                        //     )
+                        // globalclient.channels.cache.get(dataChannellog.ChannelID).send({embeds: [embedLog]});
                     };
-                    // Delete the message
-                    await channel.bulkDelete(messages, true);
-                    const embed = new EmbedBuilder()
-                        .addFields(
-                            { name: 'Cleard Messages:', value: `${embedMsg}` }
-                        )
-                    await interaction.reply({ embeds: [embed], ephemeral: true });
-                    let embedLogMsg = `${messages.length} Messages from ${stringUser} from the past Two weeks got cleared in ${channel}.`
-                    const embedLog = new EmbedBuilder()
-                        .addFields(
-                            { name: 'Cleard Messages:', value: `${embedLogMsg}` }
-                        )
-                    globalclient.channels.cache.get(dataChannellog.ChannelID).send({embeds: [embedLog]});
+                // Error Messages
+                } else {
+                    await interaction.reply({ content: `${lang.error.noadminperms}`, ephemeral: true });
                 };
             } else {
-                await interaction.reply({ content: 'This command is not available right now.', ephemeral: true });
+                await interaction.reply({ content: `${lang.error.cmdoff}`, ephemeral: true });
             };
         } else {
             console.log(`[${DateTime.utc().toFormat(timeFormat)}][ClanBot] Interaction of Command \'clear\' returned \'null / undefined\'.`);
